# Improve CRC32 performance