@Numpsy 2020-09-22
#319 did something similar with a variation of the Crc32,Net code that had some pretty nice improvements (There are couple of benchmarks in #328), but unfortunately has been sitting around for some time.
It would be nice to pick up the speed ups though.