@nsgomez 2020-09-23
We'd also love to have the perf improvement, let me know if I can help with getting either #319 or this PR merged in.