@piksel 2020-10-03
Yeah, we have some PRs for CRC improvements, but it felt kind of overwhelming to compare and choose. I made some efforts on setting up benchmarks to compare them, but we never reached a conclusion on what to do here. I will try to look in to it this weekend.